requirements.txt
mmh3==3.0.0
grpcio==1.27.1
numpy==1.21.0
pandas~=1.2.0
pyarrow==6.0.1
antlr4-python3-runtime==4.12.0
etcd3==0.12.0
awswrangler==2.18.0
boto3==1.26.27
magicdb~=1.0.7
requests~=2.28.2
botocore~=1.29.74
setuptools~=47.1.0